Modularize sections for reusability
