+++ academic_finder.py
@@ -1,9 +1,11 @@
 import requests
 import time
-import json
 
 from bs4 import BeautifulSoup
 from concurrent.futures import ThreadPoolExecutor
+
+from printing_utils import Printer
+from database_utils import load_data
 
 class Result:
   def __init__(self, title, url):
@@ -11,44 +13,6 @@ class Result:
     self.url = url
 
 results = []
-
-def load_data():
-  with open('database.json', 'r') as file:
-    return json.load(file)
-
-class Colors:
-  RED = '\033[91m'
-  GREEN = '\033[92m'
-  YELLOW = '\033[93m'
-  BLUE = '\033[94m'
-  WHITE = '\033[97m'
-  MAGENTA = '\033[35m'
-
-class MyPrinter:
-  
-  @staticmethod
-  def print_field(message):
-    print(f"{Colors.MAGENTA}{message}{Colors.MAGENTA}")
-  
-  @staticmethod
-  def print_warning(message):
-    print(f"{Colors.YELLOW}{message}{Colors.YELLOW}")
-  
-  @staticmethod 
-  def print_info(message):
-    print(f"{Colors.WHITE}{message}{Colors.WHITE}")
-  
-  @staticmethod
-  def print_finding(message):
-    print(f"{Colors.GREEN}{message}{Colors.GREEN}")
-  
-  @staticmethod
-  def print_error(message):
-    print(f'{Colors.RED}Error: {message}{Colors.RED}')
-  
-  @staticmethod
-  def print_link(message):
-    print(f'{Colors.BLUE}AT: {message}{Colors.BLUE}')
 
 def fetch_journal_data(u_name, u_url, journal, search_criteria):
   url = u_url + journal.get("url") + search_criteria
@@ -59,17 +23,17 @@ def fetch_journal_data(u_name, u_url, journal, search_criteria):
       soup = BeautifulSoup(response.text, 'lxml')
       divs = soup.find_all('div', class_='article-summary-title')
       if divs:
-        MyPrinter.print_warning(f'{name}')
-        MyPrinter.print_info(f'Results: {len(divs)}')
+        Printer.print_warning(f'{name}')
+        Printer.print_info(f'Results: {len(divs)}')
         for div in divs:
           title = div.get_text(strip=True)
           link = div.find('a')['href']
           result = Result(title, link)
           results.append(result)
-          MyPrinter.print_finding(title)
-          MyPrinter.print_link(link)
+          Printer.print_finding(title)
+          Printer.print_link(link)
     else:
-      MyPrinter.print_error(response.status_code)
+      Printer.print_error(response.status_code)
 
 def get_search_criteria():
   search_criteria = input('Please enter a search criteria without using punctuation symbols: ')
@@ -104,7 +68,8 @@ def main():
 
   end_time = time.time()
   total_time = end_time - start_time
-  MyPrinter.print_warning(f'Time: {total_time}')
+  Printer.print_warning(f'Time: {total_time}')
+  input('Press any key to exit')
 
 if __name__ == "__main__":
   main()
